# storage/engine: micro-optimize FindSplitKey