--- a/pkg/storage/engine/engine.go
+++ b/pkg/storage/engine/engine.go
@@ -96,12 +96,16 @@ type Iterator interface {
 	// the split is roughly targetSize bytes. The returned key will never be
 	// chosen from the key ranges listed in keys.NoSplitSpans and will always
 	// sort equal to or after minSplitKey.
+	//
+	// DO NOT CALL directly (except in wrapper Iterator implementations). Use the
+	// package-level MVCCFindSplitKey instead. For correct operation, the caller
+	// must set the upper bound on the iterator before calling this method.
 	FindSplitKey(start, end, minSplitKey MVCCKey, targetSize int64) (MVCCKey, error)
 	// MVCCGet is the internal implementation of the family of package-level
 	// MVCCGet functions.
 	//
-	// There is little reason to use this function directly. Use the package-level
-	// MVCCGet, or one of its variants, instead.
+	// DO NOT CALL directly (except in wrapper Iterator implementations). Use the
+	// package-level MVCCGet, or one of its variants, instead.
 	MVCCGet(
 		key roachpb.Key, timestamp hlc.Timestamp, opts MVCCGetOptions,
 	) (*roachpb.Value, *roachpb.Intent, error)
@@ -110,8 +114,10 @@ type Iterator interface {
 	// returned raw, as a buffer of varint-prefixed slices, alternating from key
 	// to value, where numKVs specifies the number of pairs in the buffer.
 	//
-	// DO NOT CALL directly. Use the package-level MVCCScan, or one of its
-	// variants, instead.
+	// DO NOT CALL directly (except in wrapper Iterator implementations). Use the
+	// package-level MVCCScan, or one of its variants, instead. For correct
+	// operation, the caller must set the lower and upper bounds on the iterator
+	// before calling this method.
 	//
 	// TODO(peter): unexport this method.
 	MVCCScan(