--- a/pkg/storage/engine/pebble_iterator.go
+++ b/pkg/storage/engine/pebble_iterator.go
@@ -239,9 +239,9 @@ func (p *pebbleIterator) ComputeStats(
 
 // Go-only version of IsValidSplitKey. Checks if the specified key is in
 // NoSplitSpans.
-func isValidSplitKey(key roachpb.Key) bool {
-	for _, noSplitSpan := range keys.NoSplitSpans {
-		if noSplitSpan.ContainsKey(key) {
+func isValidSplitKey(key roachpb.Key, noSplitSpans []roachpb.Span) bool {
+	for i := range noSplitSpans {
+		if noSplitSpans[i].ContainsKey(key) {
 			return false
 		}
 	}
@@ -252,17 +252,28 @@ func isValidSplitKey(key roachpb.Key) bool {
 func (p *pebbleIterator) FindSplitKey(
 	start, end, minSplitKey MVCCKey, targetSize int64,
 ) (MVCCKey, error) {
-	const timestampLen = int64(12)
-	p.Seek(start)
-	p.keyBuf = EncodeKeyToBuf(p.keyBuf[:0], end)
-	minSplitKeyBuf := EncodeKey(minSplitKey)
-	prevKey := make([]byte, 0)
+	const timestampLen = 12
 
 	sizeSoFar := int64(0)
 	bestDiff := int64(math.MaxInt64)
-	bestSplitKey := MVCCKey{}
+	var bestSplitKey MVCCKey
+	var prevKey []byte
 
-	for ; p.iter.Valid() && MVCCKeyCompare(p.iter.Key(), p.keyBuf) < 0; p.iter.Next() {
+	// We only have to consider no-split spans if our minimum split key possibly
+	// lies before them. Note that the no-split spans are ordered by end-key.
+	noSplitSpans := keys.NoSplitSpans
+	for i := range noSplitSpans {
+		if minSplitKey.Key.Compare(noSplitSpans[i].EndKey) <= 0 {
+			noSplitSpans = noSplitSpans[i:]
+			break
+		}
+	}
+
+	// Note that it is unnecessary to compare against "end" to decide to
+	// terminate iteration because the iterator's upper bound has already been
+	// set to end.
+	p.Seek(start)
+	for ; p.iter.Valid(); p.iter.Next() {
 		mvccKey, err := DecodeMVCCKey(p.iter.Key())
 		if err != nil {
 			return MVCCKey{}, err
@@ -272,19 +283,26 @@ func (p *pebbleIterator) FindSplitKey(
 		if diff < 0 {
 			diff = -diff
 		}
-		if diff < bestDiff && MVCCKeyCompare(p.iter.Key(), minSplitKeyBuf) >= 0 && isValidSplitKey(mvccKey.Key) {
+		if diff > bestDiff && bestSplitKey.Key != nil {
+			break
+		}
+
+		if diff < bestDiff &&
+			(minSplitKey.Key == nil || !mvccKey.Less(minSplitKey)) &&
+			(len(noSplitSpans) == 0 || isValidSplitKey(mvccKey.Key, noSplitSpans)) {
 			// We are going to have to copy bestSplitKey, since by the time we find
 			// out it's the actual best split key, the underlying slice would have
 			// changed (due to the iter.Next() call).
 			//
 			// TODO(itsbilal): Instead of copying into bestSplitKey each time,
-			// consider just calling iter.Prev() at the end to get to the same key.
+			// consider just calling iter.Prev() at the end to get to the same
+			// key. This isn't quite as easy as it sounds due to the mvccKey and
+			// noSplitSpans checks.
 			bestDiff = diff
 			bestSplitKey.Key = append(bestSplitKey.Key[:0], mvccKey.Key...)
 			bestSplitKey.Timestamp = mvccKey.Timestamp
-		}
-		if diff > bestDiff && bestSplitKey.Key != nil {
-			break
+			// Avoid comparing against minSplitKey on future iterations.
+			minSplitKey.Key = nil
 		}
 
 		sizeSoFar += int64(len(p.iter.Value()))