Both the Pebble and RocksDB implementations of `Iterator.FindSplitKey` were
performing an unnecessary comparison against the end key. This comparison
was unnecessary because `Iterator.FindSplitKey` is always called via
`MVCCFindSplitKey` which sets the upper bound of the iterator to the end
key.

Change `pebbleIterator.FindSplitKey` to avoid comparing against the
no-split spans and against the min-split key in common cases.

```
name                                      old time/op   new time/op    delta
MVCCFindSplitKey_Pebble/valueSize=32-16    79.5ms ± 3%    49.0ms ± 6%  -38.42%  (p=0.000 n=10+10)
MVCCFindSplitKey_RocksDB/valueSize=32-16   85.7ms ± 1%    82.3ms ± 1%   -3.91%  (p=0.000 n=9+9)

name                                      old speed     new speed      delta
MVCCFindSplitKey_Pebble/valueSize=32-16   844MB/s ± 3%  1372MB/s ± 5%  +62.51%  (p=0.000 n=10+10)
MVCCFindSplitKey_RocksDB/valueSize=32-16  783MB/s ± 1%   815MB/s ± 1%   +4.07%  (p=0.000 n=9+9)
```

Release note (performance improvement): Improved the performance of a
low-level routine involved in split operations.